Add --select-and-resize option and feature
This new --select-and-resize option acts like the
--select option, except that it is possible to move and
resize rectangle once it's drawn. The action may still
be canceled by pressing Escape or right-clicking. The
rectangle is accepted (screenshot is taken) when pressing
Enter.

This feature is implemented by adding a selection state,
which is, at any time, one of the following:

  * drawing a rectangle (equivalent to old `self.started`)
  * moving the rectangle
  * resizing using the top-right corner
  * resizing using the bottom-right corner
  * resizing using the bottom-left corner
  * resizing using the top-left corner
  * free to start one of the above

Bounds checking is implemented for all actions to make sure the
rectangle position/dimension do not go beyond the screen limits.